+++ backend/core/src/applications/entities/application.entity.ts
@@ -243,6 +243,7 @@ export class Application extends AbstractEntity {
   @IsBoolean({ groups: [ValidationsGroupsEnum.default] })
   markedAsDuplicate: boolean
 
+  // This is a 'virtual field' needed for CSV export
   @Expose()
   @IsBoolean({ groups: [ValidationsGroupsEnum.default] })
   @IsOptional({ groups: [ValidationsGroupsEnum.partners] })