+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file. The format 
 - Fixed:
 
   - optional fields not being marked as optional in frontend client (missing '?' indicator) ([#1470](https://github.com/bloom-housing/bloom/pull/1470))
+  - add duplicates to CSV export ([#1352](https://github.com/bloom-housing/bloom/issues/1352))
 
 - Changed:
   - User module has been removed and incorporated into Auth module